// src/theme/types.ts
export type Breakpoints = string[];

export type MediaQueries = {
  xs: string;
  sm: string;
  md: string;
  lg: string;
  xl: string;
  nav: string;
  maxWidthMD: string;
  mobNavsm: string;
  mobNavxs: string;
  maxWidthBenefitsm: string;
  colmd: string;
  colsm: string;
  MaxSm: string;
  MaxMd: string;
  Maxlg: string;

};

export type Spacing = number[];

export type Radii = {
  small: string;
  default: string;
  card: string;
  circle: string;
};

export type Shadows = {
  level1: string;
  active: string;
  success: string;
  warning: string;
  focus: string;
  inset: string;
};

export type Gradients = {
  bubblegum: string;
  cardHeader: string;
  blue: string;
  violet: string;
  violetAlt: string;
};

export type Colors = {
  primary: string;
  primaryBright: string;
  primaryDark: string;
  secondary: string;
  success: string;
  failure: string;
  warning: string;
  background: string;
  backgroundDisabled: string;
  backgroundAlt: string;
  text: string;
  textDisabled: string;
  textSubtle: string;
  borderColor: string;
  aboutUsBackground: string;
  artistsBackground: string;
  navbackground: string;
  alertBackgroundColor : string;
  blackWhite : string;
  activeMobileNav: string;
  cartActionBg :string;
  whiteBlack:string
};

export type ZIndices = {
  dropdown: number;
  modal: number;
};